backend/requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
backoff==2.2.1
backports.zstd==1.0.0
beautifulsoup4==4.14.2
Brotli==1.1.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
clerk-sdk==0.3.7
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
environs==14.3.0
fastapi==0.120.1
fastapi_cors==0.0.6
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.0
google-api-python-client==2.185.0
google-auth==2.42.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
marshmallow==4.0.1
multidict==6.7.0
packaging==25.0
postgrest==2.22.3
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.11.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
realtime==2.22.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.1
storage3==2.22.3
StrEnum==0.4.15
supabase==2.22.3
supabase-auth==2.22.3
supabase-functions==2.22.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
